have stream generate ValidationOutcome
